added success rate for quizzes; switched success rate type for wordpairs to double; started working on quiz progress merger

+++ SteelQuiz/QuizProgressData/WordProgData.cs
@@ -58,7 +58,7 @@ namespace SteelQuiz.QuizProgressData
             return WordTries.Count;
         }
 
-        public int GetSuccessRate()
+        public double GetSuccessRate()
         {
             var tries = GetWordTriesCount();
             if (tries == 0)
